+++ src/perflogger/perflogger.py
@@ -2,6 +2,8 @@ import inspect
 import time
 
 from os import PathLike
+from pathlib import Path
+import jsonlines
 
 from .logger import Logger
 
@@ -22,6 +24,7 @@ class PerfLogger(Logger):
         )
         self.tick = tick_func
         self.avgs_step = avgs_step
+        self.output_file_path = Path(output_file_path) if output_file_path else None
 
         self.tags = []
         self.tags_data = {}
@@ -41,7 +44,10 @@ class PerfLogger(Logger):
         """Add a new tag in the tags table as the delta between now and last tag;
         append that delta to the tags_data table to later compute averages"""
         now = self.tick()
-        self.tags_data.setdefault(tag_name, []).append(now - self.tags[-1][1])
+        delta = now - self.tags[-1][1]
+        if self.output_file_path is not None:
+            self.write_to_file(tag_name, now, delta)
+        self.tags_data.setdefault(tag_name, []).append(delta)
         self.tags.append((tag_name, now))
 
     def _set_averages(self):
@@ -49,6 +55,21 @@ class PerfLogger(Logger):
         self.avgs.clear()
         for name, data in self.tags_data.items():
             self.avgs[name] = sum(data) / len(data)
+
+    def _write_to_file(self, lines):
+        with self._get_write_lock(self.output_file_path):
+            with jsonlines.open(self.output_file_path, "a") as file:
+                file.write(lines)
+
+    def write_to_file(self, tag_name, value, delta):
+        line = {
+            "logger_name": self.name,
+            "logger_class": type(self).__name__,
+            "value_name": tag_name,
+            "value_plain": value,
+            "value_delta": delta,
+        }
+        self._write_to_file(line)
 
     def get_deltas(self):
         """Return deltas stored in the tags table both in a human readable format and true float"""